--- aidial_sdk/pydantic/__init__.py
@@ -1,2 +1,0 @@
-from ._compat import PYDANTIC_V2, ConfigDict
-from ._models import BaseModel, model_validator
